(PAR): implement `no_struct_literal_expr`
When parsing condition in `if/when/match`, certain kinds of
expressions are forbidden. Namely, struct literals (`Foo {i: 92}`) and
block-ending ranges (`1..{42}`). This helps in such situations:

```
if i < Foo {
   ...
}

for x in 1.. {
    ...
}

```

This commit adds a boolean flag to the parser, which keeps track of the
context and allows to impose the restrictions without duplicating the
grammar for expressions. Rustc's `libsyntax/parser/parser.rs` handles
the situation in a similar way.

Reference: https://github.com/rust-lang/rfcs/blob/master/text/0092-struct-grammar.md