(RES): implement child modules resolve
When we see `mod foo;` we should try to find a file corresponding to
`foo`. It is possible only if a parent module is a directory owner.

A module is a directory owner if it is either
  * a crate root
  * a nested inline module
  * called `mod.rs`

It is an error to have `mod foo;` in a non directory owning module.

The file name for `foo` is (modulo #[path = ] attribute) a directory
name of the parent module plus either `/foo.rs` or `/foo/mod.rs`.

Example 1

```Rust
// file src/main.rs, a crate root

mod foo; // src/foo.rs

mod inner {
    mod foo; // src/inner/foo.rs
}
```

Example 2

```Rust
// file src/foo.rs

mod bar; // forbidden

mod inner {
    mod foo; // src/inner/foo.rs
}
```

The semantics of `mod foo;` nested inside a function is not fully
specified at the moment: https://github.com/rust-lang/rust/issues/29765