(RES): fix crossmodule resolve during completion
For autocomlete IDEA inserts dummy "IntellijIdeaRulezzz" identifier at
the caret position, and calls `getReference` on the resulting element.
To insert this identifier, IDEA makes a copy of the PSI file. So during
autocomplete the file is non-physical, it's virtualFile is `null` and
resolve of parent and children modules does not work.

Luckily there is `getOriginalFile` method which we should really employ
for the purpose of resolve.