(RES): relax type bounds in resolve
Resolve to `RustCompositeElement` instead of `RustNamedElement`.

Resolving to `NamedElement` is both useless and wrong.

It is useless because element can be renamed several times via `as`
alias, so the name of the resolved element has no connection with the
name of the referring element.

It is wrong because in fact it is possible to resolve to an anonymous
element!

Example 1, tuple struct fields:

```
struct Pair(String, i32);

fn main() {
    let p: Pair = Default::default();
    p.<caret>1; // this should resolve to the anonymous field of Pair.
}
```

Example 2, the root module is always anonymous so `::` resolves to a
nameless item.