Add documenting for trait members.
Also adds tests and adjusts parse for trait type members slightly.
This is because type parameters are considered an independent
named element, and it did not make grammatical sense to have
that rule be used in the place where associated types are declared.
The tests checking for this have been adjusted accordingly.