(RES): refactor DeclaringElements
It turned out that a significant part of DeclaringElments had a trivial
`declarations` implementation that just returns a singleton
`listOf(this)`. In fact, only patterns really took advantage of being
DeclaringElements.

Let's simplify the hierarchy to this

`ResolveScope` determines tree walking up traversal during resolve
and must supply a collection of `NamedElements`, defined in the scope.

`DeclaringElement` is usually a child of `ResovleScope`. It can list
`NamedElements`, but does not participate in the tree traversal itself.
Instead, parent `ResolveScope` delegates to the `DeclaringElement`.

`NamedElement` is the potential target of the resolve.