(MISC): kill buggy `recursionGuard`
`RecursionManager.doPreventingRecursion` relays on the "identity" of the
`Computable` passed in. That is, `computable.class.name` is
used as a key for identifying a particular computation. So it is
important that each invocation of `doPreventingRecursion` uses a fresh
class for the `Computable`. And this is not the case with this
`recursionGuard` function because SAM conversion of `() -> T` into
`Computable<T>` happens *once* inside the helper.