(T): don't invalidate stub indices in resolve tests
This invalidation simply didn't work before 2016.2. In 2016.2 it works
and increases testing time from 30 seconds to 5 minutes.

The invalidation is not necessary because indexing should not depend on
any external state. Put it another way, if invalidation changes the
observed behavior then we have a serious bug with our indexing.