(ANN): Do not report an error when declaring a module with an explicit path
Normally it is not allowed to declare a new module if the current file
does not own the current directory.
If the module declaration has an explicit path attibute
it is allowed anyway.

main.rs
mod module1;

module1.rs:
[path="./test.rs"]
mod test; //Wrong error: "Cannot declare a new module at this location"