(TYP): Fix integer/float primitive type parsing and literal type detection
There are different string comparison requiremens for parsing a literal
and for parsing a type.
When parsing the type we need to check the whole type name.
When parsing a literal we just need to check if the postfix matches.

struct myu32;
type T = myu32;

"myu32" was erroneously highlighted as a keyword/primitive
since it ended with "u32".

let num1 = 1.0f32;
let num2 = 1.0f64;

The type of the floating point literals did not depend on the f32/f64 postfix.
Instead it always was the type f32.