Better desicion making on statement vs expression.
By checking for wether this expression is top level in the block,
so it's a whole expression on a line but it doesn't have a semicolon at the end.
we correctly introduce the let on the same line.

Further more we also get rid of path elements in possible expressions,
they don't make sense to put in a local variable