CARGO: kill cargo processes when a module is closed
This fixes a long-standing CI problem with `ThreadLeaked` exceptions.

This also ameliorates a spectacular bug in Cargo:

https://github.com/rust-lang/cargo/issues/3564

If the internet goes down while Cargo updates the index, it may get stuck in a
select with infinite timeout inside of `libgit2`, holding an exclusive lock on
the index and preventing all other Cargo instances from running.

Lessons learned: if the test fails sometimes during CI, it is probably trying to
say something important.