CI: compatibility between Grammar Kit and EAP
So, we support 2016.2, 2016.3 and 2017-EAP. GK 1.4.0 is compatible with 2016.2
and 2016.3. GK 1.5.0 (currently release) is compatible with 2016.3 only. Current
master of GK works with 2016.2 and EAP.

But we need GK only during code generation, and the compatibility of GK does not
matter much as long as the **generated code** works fine.

Let's run GK with a fixed version of idea!