DBG: first scratches
The hardest thing about debugger is how to actually get to the infrastructure
from CLion. There isn't a good way to do this yet, so currently debugger is
segregated into completely separate plugin in `/debugger`. It is not build
during the usual CI, because building is tricky an works only on Linux:

```
cd debugger
./download_deps.sh
../gradlew :runIdea
```

Nevertheless, you can install this plugin in CLion and get step into and local
variables working. Needless to say, this is **very** unstable yet, so installing
debugger in theory can harm your C++ debugging experience.