TY: Check mutability when comparing/unifying reference/pointer types
Fixes #1667
where the wrong mutability would be inferred for the following code:

let mut b: Vec<Vec<usize>> = vec![vec![1, 2]];
for a in &mut b {
    a
  //^ &mut Vec<usize>
}

IntoIterator is implemented for &Vec<T> and also &mut Vec<T>

To typify the 'for' loop, ImplLookup.findIteratorItemType() searches
for the first implementation of the Iterator or IntoIterator trait.

Since TyReference.unifyWith() did not differentiate between & and &mut references
the IntoIterator implementation with the wrong mutability would be accepted
resulting in the wrong mutability of the iterated item.

The same problem also applies to pointers.
There it mostly causes resolving of inherent functions on pointers to fail
because multiple implementations are found.