TY: Move detection of result::Result<T, E> to LangItems.kt
The type Result<T, E> is a well known type in Rust an important for
autocompletion of the ? expression and the try! macro.
Because of this its detection should be done in a central place.

The current code is still mostly an approximation but it can be shared
by the rest of the application.