TY: Improve resolving of primitive::method() calls
This mostly helps for char, str, f32 and f64 types.
For integer types it does not help because on them all inherent impls
are done via macros.
For pointers it does not help because for them it is necessary
to use UFCS syntax which is not resolved yet.
<*const char>::is_null(p);
<char>::is_digit('1', 10)

Primitive types in the first part of a call
are now also highlighted as such.

char::is_lowercase('a');
^ formatted as a primitive type

This unfortunately also applies for times when a primitive
has the same name as a module

use std::f64;
f64::EPSILON;
^ formatted as a primitive type
while it is actually refers to module name in this case