ANN: Extract the check for surrounding unsafe block or fn
Extract a function so it can be reused.

While doing this, fix a mistake in the original implementation.

The find { } looks for a RsBlockExpr but the if() checks for a RsBlock.
This caused the left part of the if() to always evaluate to false.
Luckily this had no impact on the correctness as the code
is only interested in blocks or functions missing the unsafe keyword.