Switch to additional library roots provider
Before, we used to create a *persistent* IDEA library for stdlib and
crates from crates IO. Starting from 2017.1, it became possible to
calculate them on the fly, which makes more sense for our use case.

Note however that now we can't associate a psi element from libraries
with modules. And this is probably a good thing, because different
modules may reuse the same stdlib anyway.