Fix exceptions after running the "Change try! to ?" inspection
A try!() invocation looks like this in PSI:
RsMacroExpr
-RsMacroCall
--identifier
--!
--RsTryMacroArgument

RsTryMacroInspection registers the problem and the fix on the RsMacroCall.
When the fix is applied, it replaces the RsMacroCall node with a RsTryExpr,
leaving back a corrupt RsMacroExpr missing a required RsMacroCall child.

Later some other code trips over this corrupt PSI node
triggering an exception in PsiElementBase.notNullChild().

This problem was introduced by commit 2e25378cc0
which reorganized macro related grammar.

Fix the problem by visiting and working on the RsMacroExpr instead.