ANN: Coerce *mut pointers to *const pointers to fix a wrong E0308
In places that expect a *const T it is possible to pass a *mut T
due to coercion.
See https://doc.rust-lang.org/nomicon/coercions.html

This fixes a wrong E0308 error when using a *mut T at a points where
a *const T is expected.
(function argument, let statement, ...)

CC: #1753