ANN: Handle unsafe impl of safe traits required by #[may_dangle] attribute
Since the implementation of RFC 1327
https://github.com/rust-lang/rfcs/blob/master/text/1327-dropck-param-eyepatch.md
some implementations of the "Drop" trait in the std library
would be marked with the error E0199.

The presence of a #[may_dangle] attribute on a type parameter
or a lifetime parameter requires an unsafe impl of the safe "Drop" trait
which we did not handle until now.

unsafe impl<#[may_dangle] T> Drop for Vec<T> {
    fn drop(&mut self) {
        ...
    }
}