CARGO: Refresh the cargo project when src/lib.rs is created or deleted
Due to a typo we would not react to src/lib.rs being created.
This would cause us to not refresh our cargo workspace / crate roots.
(WorkspaceImpl.targetByCrateRootUrl)

This in turn could cause a wrong error
"Cannot declare a new module at this location"
because src/lib.rs would not be recognized as crate root.

Fixes #1638