CARGO: Make CargoTomlWatcher a bit more strict about which files it accepts
The code mainly uses String.endsWith() which will also match
when there were is a prefix before the directory/file names.

Slightly improve this by adding path separators to the beginning
of the patterns.

The same problem can still happen deeper in the project hierarchy.
But since it just causes us to refresh the cargo workspace
when such a file is created it is not too terrible.

Also since we do not check the start of the path we can be triggered
accidentally when Cargo.toml inside a different project changes.