RES: Fix handling of #[macro_export] and #[macro_use] attributes
The #[macro_export] attribute is necessary to export a macro
to the outside of a crate.
To use a macro inside the same crate (even in different module)
this attribute is not necessary.

The #[macro_use] attribute is necessary to make a macro available
to the parent module inside the same crate.
For exporting a macro this attribute is not necessary.

See https://doc.rust-lang.org/reference/attributes.html

Remove a few unnecessary #[macro_use] and #[macro_export] from tests.