RES: Fix resolve inside modules specified via explicit #[path = "foo/mod.rs"]
The path attribute allows to load the module source from a file
with a different name/path than the module.

When the explicit path ends with "/mod.rs" then resolving
would not work properly inside this module.

This is because a wrong key (always named "mod")
would be added to RsModulesIndex for #[path = "foo/mod.rs"].
This would cause RsMod.`super` to fail.
And this in turn would cause getting the crate root to fail.

This improves resolving in libstd/sys/mod.rs