GRAM: Do not throw exceptions when outer attributes appear on literals
RsLitExpr.kind accesses firstChild of a RsLitExpr to check its literal type.

Since #2039 we allow outer attributes as part of a RsLitExpr
which means that accessing firstChild can yield a literal or a outer attribute.
When we get an outer attributes we would not correctly identify
the literal type and throw an exception.

Instead use ASTNode.findChildByType() to find the correct child AST node.

Fixes #1893