WIP: NAV: Implement Ctrl-Alt-B (Go to implementation) for const, fn, type
To find implementations of a RsAbstractable we first search
for implementations of the trait and then use mapQuery() to
find implementations of the RsAbstractable.

This requires us to also use mapQuery() on existing searches for
implementations of trait, struct and enum sow the return type
of the query matches.

TODO:
- There are currently no tests for searching for implementations