TY: Support setting the enum discriminant type via #[repr(u*)] attribute
For enum types it is possible to change the type/size
of the discriminant via a #[repr()] attribute.
Rust allows using any u* or i* type like u8, i16, ...
This is mainly used in FFI code.

Correctly typify the discriminant type when this attribute is present
instead of using the default type of isize.

Fixes #2477

(cherry picked from commit 9a795bd)