ANN: add type mismatch quick fix using Borrow/BorrowMut and AsRef/Mut
This commit adds series of quick-fix for the types mismatch when
impl `Borrow<B>`/`BorrowMut<B>/AsRef<B>/AsMut<B>` for A exists
where A is actual type and B is expected type,
which fulfills part of the forth bullet-point of the issue:
https://github.com/intellij-rust/intellij-rust/issues/1730.

Also refactoring a couple of existing quick-fixes of the similar form
to reduce code duplication a little bit, and to make keeping the
messaging/naming unified easier.

Also fixing ConvertToOwnedTyFixTest.

(cherry picked from commit 97db106)