INSP: don't highlight paths with primitive type name as unresolved.
Now we ignore primitive types in a path if all path is a primitive type name.
But when primitive type name is part of path, for example `f64::abs`,
we highlight it as an unresolved reference because our resolve can't resolve primitive types.

This fix checks if base path is a primitive type name instead of full path

(cherry picked from commit 87ecf5d)