TY: Infer type of macros like cfg!() column!(), module_path!(), ...
This macros are built into the compiler and the return types
are not visible from the macro definition in stdlib.
Thus we have to hardcode the return type.

For include_bytes!() we infer the byte array size to be unknown.
If necessary we could check the actual file size, as we can already
resolve the path to the file.