PERF: Added several `ProgressManager.checkCanceled()` calls
`ProgressManager.checkCanceled()` is used to cancel (by throwing an
an exception) the current task. I.e. if user requested completion
and then canceled it. This call is often found in the platform (e.g.
in ResolveCache we call `checkCanceled()` on each cache lookup),
so we can decide that it is cheap and its use is good practice.

In this commit I've added `checkCanceled()` to some possible
long recursive/loop operations like type inference.