CARGO: On double click jump from Cargo tool window to Cargo.toml
In the Cargo tool windows there are certain points for which
it is nice to jump to Cargo.toml on double clicks
- a workspace member
- a package
- a feature in a package

Currently we just jump to Cargo.toml keeping the current cursor/selection.
In the future we might want to jump to the exact location
of a feature or the [features] section.