TY: Add more tests for match ergonomics around `mut` and `ref` keywords
Add tests for various combinations of `mut` and `ref` keywords
in the short and long form.

This tests show some problems around handling of `mut` and `ref` in
struct/enum patterns.

Link to the match ergonomics RFC from the tests, so it is easier
to check what is tested here.

CC #3225