Add external package with the same workspace
There is already `dep-lib` for separation code into different crates,
but it has PackageOrigin.DEPENDENCY, thus it is treated like external
dependency which we cannot modify. However, sometimes it is useful to
create tests that rely on different crates, but the same workspace. For
example, there can be some fixes that cannot modify external
dependencies, but have to modify any code that is located within one
workspace, even in different crates.