PERF&MACRO: Introduce MacroExpansionFileSystem
Expanded macros now are not stored on a disk as
individual physical files. Instead, they are
stored in a snapshot only (snapshot is an
IntelliJ platform concept that means a copy
of a part of filesystem stored in embedded
DB in IntelliJ directory) with new
`MacroExpansionVFS`.

Now macro expansion should occupy much less
disk space. Also, this should speed up macro
expansion because of fewer system calls
(filesystem interaction).