Fix LLDB pretty-printer for Rc/Arc on 2020.1
Since 2020.1 our LLDB contains Rust type system so `SBType.GetBasicType()`
fails when called on Rust type supposedly because `eBasicTypeLong` is C type,
which may be incompatible with Rust `SBType` somehow.
In contrast, `SBTarget.GetBasicType()` uses only `ClangASTContext`
to compute the type and doesn't depend on Rust `SBType`,
so it works fine in that case.