Drop processItemDeclarationsWithCache
This cache was introduced before the latest
optimization series and now is not actual.

I measured the impact of its removal and it
is zero or even positive (some benchmarks
works faster without the cache, lol)

But the main reason for its removal is
(rare) `Non-idempotent computation` assert
triggering. Unfortunately I didnt found
the real reason of this