MACRO: process macro calls better
The previous version of macro expansion engine processed all project files on the first step.
It led to a situation when a module was added to the module structure via macro call and at the same time, it had own macro calls.
It such cases we failed to expand module's macro calls because name resolution for macro calls didn't work properly.

These changes introduce pending files - project files that are not belong to the module tree.
Such files are postponed to expand their macro calls until they belong to module tree.
It should solve "File is not included into module tree" issue for some popular crates like `async-std` and `tokio`