#5165: Provide builtin Clone & Copy impls for TyFunction
Previously, `TyFunction` was considered as non-Copy & non-Clone type in terms of our type inference.
Although there was a special case in `Ty.isMovesByDefault` to handle `TyFunction` properly during move analysis,
it did not work in case of `TyFunction` as a type parameter.
Now `TyFunction` is considered as Copy & Clone type so the special case in `Ty.isMovesByDefault` has been removed.