GRD: enable Gradle Build Cache
From [Gradle docs](https://docs.gradle.org/current/userguide/build_cache.html):

> The Gradle build cache is a cache mechanism that aims to save time by reusing
> outputs produced by other builds. The build cache works by storing build outputs
> and allowing builds to fetch these outputs from the cache when it is determined
> that inputs have not changed, avoiding the expensive work of regenerating them.

In short, now we have a directory `build/build-cache` where Gradle puts all
(intermediate) build outputs it ever produces and reuses them if possible,
so frequent git branch switching or IntelliJ platform version switching
does not cause long build times.

In order to prevent constant growth of `build/build-cache` there is a mechanism that
removes cached entries that have not been accessed for 30 days (the value is
configured in `settings.gradle.kts`). My `build-cache` directory size is less than
150 Mb after a week of usage.

Now you can mention these messages in a build log:

```
> Task :compileKotlin FROM-CACHE
> Task :compileJava FROM-CACHE
```

Such `FROM-CACHE` records (instead of the usual `UP-TO-DATE`) means that the task
was not `UP-TO-DATE`, but it was not re-launched. Instead, the build output
was fetched from the build cache.

For me, this drastically improves the build time in the case of the platform version
switching.