Don't throw exception on invalid user-created paths
During toolchain detection, the plugin checks `PATH` variable and tries to convert its parts to `java.nio.file.Path`.
But if a user appended something illegal into the variable, the conversion fails with `InvalidPathException` exception.
Previously, we did nothing with it and it broke the UI. Now, we log such cases and skip illegal paths