LI: fixed intentions and fixes running within doctests
Previously modification of injected doctest code was incorrectly handled, so that applying an fix or intention could result in corrupted code following the doctest. The reason was that modifications within a doctest run in their own injected editor, which wasn't correctly projected back into the containing file editor, but the offsets of changes were calculated relative to the whole file.