REF: fix extract variable on RsExprStmt and last expr in block
This involves a change to the tests. The old behaviour was breaking code since it changed the returned value of the block from T to (), and breaking code is bad. It's also trivially easy to change the returned value after refactoring to a more complex expression, since only a single identifier should be returned.

This commit also includes a fix for extracting several occurrences of an expression. Previously if we extracted an RsExprStmt, then the logic would short-circuit and extract only it, ignoring the possibility of other occurrences.