INSP: remove spurious unreachable warnings for match exprs with no arms
We add a dummy CFG edge from the discriminant of the match to its output in this case, since adding a match arm with basically any return type would make the above CFG path possible. This avoids spurious "unreachable expr/stmt" warnings when the user types a new match block.

If the user adds at least one explicit arm, then the dataflow will be analysed properly (it's possible to add first a noreturn arm, while the other arms would be expected to return a value, but this is an easily understandable rare edge case which I don't think requires any special treatment).

The only case when a match with no arms is valid and indeed signifies unreachability is when the discriminant has the type with no values (e.g. ! or `enum Foo {}`), but this case can only be properly handled after type inference, not at the CFG level. It is also a very rare case which shouldn't cause problems for the common case of simply typing out the match expression.