REF: fix Extract Variable interaction with closures
Previously refactoring didn't consider whether the extracted subexpression of a closure is also used somewhere outside of the closure. This lead to the binding created within the scope of the closure, while the occurrences were replaced regardless of the scope. We move the binding inside of the closure body only if all extracted expressions are within its scope.