WIP: Add checkbox to enable Rust support in MSVC LLDB
Add a checkbox in debugger settings to enable `LLDB_USE_RUST_TYPE_SYSTEM` environment variable. The checkbox is disabled by default. When the checkbox is on, MSVC LLDB tries to parse Rust PDB files and user expressions via Rust-specific classes (type system, PDB parser, user expression parser, and so on). When it is off, MSVC LLDB uses Clang classes instead, as it was before.