DBG: Add registry flag to enable native Rust support in MSVC LLDB
Add registry flag to enable native Rust support in MSVC LLDB on Windows. The flag is disabled by default and operates as follows:
- when disabled, MSVC LLDB uses Clang-based implementation and treats Rust types and expressions as C++ types and expressions (same as it was before);
- when enabled, MSVC LLDB detects if a PDB file corresponds to Rust source code, and in that case, the new Rust-aware implementation is used to manipulate debug information, types and expressions.

This functionality is available for MSVC LLDB bundled with CLion 2022.2 since 222.2889 build. Note, the Rust-aware implementation at this point is experimental and might cause LLDB crashes in certain cases.