Fix `Undo` (Ctrl+Z) action after typing in a macro call
The bug was introduced in 74d85146ac4a80f086471cbb75e21429704f4f8f.
It turned out that undo merger stopped working when typing in a macro.
That is, if you type, say, `foo` outside a macro call and invoke `Undo` action, the entire `foo` is undone.
But if you type it in a macro call, then it undoes only the last typed symbol - `o`, so `fo` remains in the document