add inspection understanding of #[derivative(...)]
IntelliJ can now infer when a value implements traits from #[derivative(...)], even if proc-macro support is turned off. For example, if the value implements #[derivative(Clone, Copy)], IntelliJ will not report move errors.

IntelliJ can also infer a limited set of custom bounds. e.g. with

```
#[derive(Derivative)]
#[derivative(Clone(bounds = ""), Copy(bounds = "")]
struct S<T>(PhantomData<T>);
```

IntelliJ will not report move errors even if `T` does not implement `Clone` or `Copy`. Only a limited syntax of custom bounds are supported, any bounds outside of this syntax are ignored.

I added test cases to verify that inspections with Clone and Copy work. I also tested it in my IDE a little.

I have tested with proc macros enabled and there didn't seem to be issues, but I'm not sure if proc macros conflict with this, since they may also infer derivative. Regardless this is still useful as proc macros are not stable yet, and even then this hardcoded analysis is bound to be faster.

Also note that the analysis doesn't actually check for derive(Derivative) or if derivative::Derivative is even in scope, it simply checks for the #[derivative(...)] annotation.