ANN: detect E0317 missing else block
Also slightly change type inference for if expressions without else block.
Previously such expressions were always inferred to `()`. This is right for expression statements,
but wrong for nested expressions. We change it to infer `()` only if the parent is a block, otherwise
type inference returns the type of the `if` block.