Introduce `TyPlaceholder`
The idea is that `TyInfer` must not be stored in a cache because it is mutable.
Previously we replaced a `TyInfer` with a fresh one after retrieving it from the cache,
but this approach is a foot-shooter since it's very easy to forget to replace it, and
then everything will continue to work as expected in must cases (most important, in
unit tests), but will shoot in 0.1% cases.
Now we replace `TyInfer` to `TyPlaceholder` when storing it to the cache, so if we
forget to replace it, it will 100% fail (even in tests!) so we will notice and fix it