GRD: use `antlr` plugin only in `debugger` module
In #9349 the plugin started to use `antlr` Gradle plugin to generate parser for debugger. But the corresponding changes used `plugin { antlr }` construction at the top-level of gradle script (it can be used only at top level) that led to application of `antlr` plugin for root module. But we need the plugin only in `debugger` module

These changes drop usage of `antlr` plugin from root module. As a result:
- no more unexpected `generateGrammarSource` task in root module
- we don't need to exclude antlr compilation time dependencies in root module
- no more implicit task dependencies in unexpected places. In expected places, they are declared explicitly more properly now