PERF: speed up syntax highlighting in macro calls
Previously we highlighted macro expansion and then mapped annotations
to a macro call body. It was slow due to several reasons:
1. macro expansions are usually larger than macro calls (in terms of
   tokens);
2. In general, Annotators are more lazy - they first highlight
   a visible part of the code and then the rest of the file.
   On the opposite, `RsMacroExpansionHighlightingPass` is not lazy -
   it highlights the entire file and only then shows highlighting.

Now we do the same for most of the annotators except
`RsHighlightingAnnotator` which now uses a reverse approach -
it maps a token from a macro call to a macro expansion and
then figures out how they should be highlighted there.