Create Rust project settings panel only if needed
Previously, we created its panel in constructor with registration of all components which require disposable object. And in common case, all UI resources are disposed in `org.rust.cargo.project.configurable.RsProjectConfigurable.disposeUIResources` and everything is fine

But for some reason, `RsProjectConfigurable` object is created during closing settings window. And for this object, platform doesn't call `disposeUIResources`. As a result, UI panel created in constructor is not disposed

Now, panel is created only if it's really needed